+++ app/javascript/packs/application.js
@@ -4,10 +4,11 @@
 // that code so it'll be compiled.
 
 require("@rails/ujs").start()
-require("turbolinks").start()
+//require("turbolinks").start()
 require("@rails/activestorage").start()
 require("channels")
 require("../sales_fee")
+require("../card")
 
 
 // Uncomment to copy all static images under ../images to the output folder and reference
